requirements.txt
requests
python-dateutil
pyinstaller
